requirements.txt
asgiref==3.7.2
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
coverage==7.4.4
decouple==0.0.7
dj-database-url==2.1.0
Django==5.0.3
django-crispy-forms==2.1
django-formtools==2.5.1
django-heroku==0.3.1
gunicorn==21.2.0
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PySocks==1.7.1
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
waitress==3.0.0
Werkzeug==3.0.1
whitenoise==6.6.0
wsproto==1.2.0
Django-Verify-Email == 2.0.3
